better split of code, use local inference
